refactor(appservice)!: Improve API and cleanup docs
